@@ -3,16 +3,16 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@reduxjs/toolkit": "^1.8.0",
+    "@reduxjs/toolkit": "^2.7.0",
     "@testing-library/jest-dom": "^5.15.1",
     "@testing-library/react": "^11.2.7",
     "@testing-library/user-event": "^12.8.3",
     "bootstrap": "^5.3.5",
-    "react": "^17.0.2",
+    "react": "^19.1.0",
     "react-bootstrap": "^2.10.9",
-    "react-dom": "^17.0.2",
+    "react-dom": "^19.1.0",
     "react-icons": "^5.5.0",
-    "react-redux": "^7.2.6",
+    "react-redux": "^9.2.0",
     "react-router": "^6.0.2",
     "react-router-dom": "^6.0.2",
     "react-scripts": "4.0.3",
